created component for UserContext
